Fixed type of SogliaSpedizioneGratuita

+++ primo-sito-web-PURE/src/main/java/it/polimi/tiw/DAO/DAO_Fornitore.java
@@ -96,7 +96,7 @@ public class DAO_Fornitore {
         while( resultSet.next() ){
         	
         	// gli attributi che possono essere null vanno trattati separatamente
-            Integer sogliaSpedizione = resultSet.getInt("SogliaSpedizioneGratuita");
+            Double sogliaSpedizione = resultSet.getDouble("SogliaSpedizioneGratuita");
             if( resultSet.wasNull() )
                 sogliaSpedizione = null;
